+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 
 ## [Unreleased]
 
+## [v0.2.3] - 2025-12-06
+
+### Added
+
+- Make `Save-NtfyAuthentication` public.
+    - Added aliases - `sva`, `Save-NtfyAuth`
+
 ## [v0.2.2] - 2025-12-06
 
 ### Added